FrontEnd/requirements.txt
flask
Pillow
tensorflow